Update pause display, now is linked with ESC key
The pause displays when you pause the game via ESCAPE key
it displays when you pause the game, and goes away when
you resume the game.  The continue button doesn't work
for now and also the pause display is not resizable
at the moment as well.  At least the exit button on the
pause display works.  The color and design for the pause
screen is new skin (Gdx.files.internal(flat-earth/skin/flat-earth-ui.json))