New EntityEffectsComponent Component
EntityEffects is a component, given to obstacles.
When a target entity collides into this object, a specified action is
performed on the target entity.

Example is the cobweb, slows the player if they walk on it.

Actions are not yet implemented, due to significant merge conflicts and
lack of implementation in the Player and Enemies Controllers.